requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.7.4.post0; python_version >= '3.6'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
discord-py-slash-command==2.3.2
discord.py==1.7.3
idna==3.2; python_version >= '3.5'
multidict==5.1.0; python_version >= '3.6'
python-dotenv==0.18.0
typing-extensions==3.10.0.0
yarl==1.6.3; python_version >= '3.6'
